+++ R/list.R
@@ -2,7 +2,7 @@
 #'
 #' List Quarto Extensions in this folder or project by running `quarto list`
 #'
-#' @inheritParams quarto_render
+#' @return A data frame with the installed extensions or NULL (invisibly) if no extensions are installed.
 #'
 #' @examples
 #' \dontrun{
@@ -10,27 +10,25 @@
 #' quarto_list_extensions()
 #' }
 #'
-#' @importFrom rlang is_interactive
-#' @importFrom cli cli_abort
-#' @importFrom utils read.table
 #' @export
-quarto_list_extensions <- function(quiet = FALSE, quarto_args = NULL) {
+quarto_list_extensions <- function() {
   quarto_bin <- find_quarto()
 
-  args <- c("extensions", if (quiet) cli_arg_quiet(), quarto_args)
-  x <- quarto_list(args, quarto_bin = quarto_bin, echo = TRUE)
+  # quarto list extensions --quiet will return nothing so we need to prevent that.
+  args <- c("extensions")
+  x <- quarto_list(args, quarto_bin = quarto_bin, echo = FALSE)
   # Clean the stderr output to remove extra spaces and ensure consistent formatting
   stderr_cleaned <- gsub("\\s+$", "", x$stderr)
   if (grepl("No extensions are installed", stderr_cleaned)) {
     invisible()
   } else {
-    invisible(utils::read.table(
+    utils::read.table(
       text = stderr_cleaned,
       header = TRUE,
       fill = TRUE,
       sep = "",
       stringsAsFactors = FALSE
-    ))
+    )
   }
 }