+++ tests/testthat/test-list.R
@@ -1,4 +1,6 @@
 test_that("Listing extensions", {
+  # don't try to install extensions on CRAN
+  skip_on_cran()
   skip_if_no_quarto()
   skip_if_offline("github.com")
   qmd <- local_qmd_file(c("content"))
@@ -6,8 +8,8 @@ test_that("Listing extensions", {
   expect_null(quarto_list_extensions())
   quarto_add_extension("quarto-ext/fontawesome", no_prompt = TRUE, quiet = TRUE)
   expect_true(dir.exists("_extensions/quarto-ext/fontawesome"))
-  expect_equal(nrow(quarto_list_extensions()), 1)
+  expect_snapshot(quarto_list_extensions())
   quarto_add_extension("quarto-ext/lightbox", no_prompt = TRUE, quiet = TRUE)
   expect_true(dir.exists("_extensions/quarto-ext/lightbox"))
-  expect_equal(nrow(quarto_list_extensions()), 2)
+  expect_snapshot(quarto_list_extensions())
 })